+++ front/lib/api/assistant/visualization.ts
@@ -1,3 +1,89 @@
+import type { ContentFragmentType, ConversationType } from "@dust-tt/types";
+import { isContentFragmentType, removeNulls } from "@dust-tt/types";
+import _ from "lodash";
+import * as readline from "readline"; // Add this line
+import type { Readable } from "stream";
+
+import type { Authenticator } from "@app/lib/auth";
+import { FileResource } from "@app/lib/resources/file_resource";
+
+export async function getVisualizationPrompt({
+  auth,
+  conversation,
+}: {
+  auth: Authenticator;
+  conversation: ConversationType;
+}) {
+  const readFirstFiveLines = (inputStream: Readable): Promise<string[]> => {
+    return new Promise((resolve, reject) => {
+      const rl: readline.Interface = readline.createInterface({
+        input: inputStream,
+        crlfDelay: Infinity,
+      });
+
+      let lineCount: number = 0;
+      const lines: string[] = [];
+
+      rl.on("line", (line: string) => {
+        lines.push(line);
+        lineCount++;
+        if (lineCount === 5) {
+          rl.close();
+        }
+      });
+
+      rl.on("close", () => {
+        resolve(lines);
+      });
+
+      rl.on("error", (err: Error) => {
+        reject(err);
+      });
+    });
+  };
+
+  const contentFragmentMessages: Array<ContentFragmentType> = [];
+  for (const m of conversation.content.flat(1)) {
+    if (isContentFragmentType(m)) {
+      contentFragmentMessages.push(m);
+    }
+  }
+  const contentFragmentFileBySid = _.keyBy(
+    await FileResource.fetchByIds(
+      auth,
+      removeNulls(contentFragmentMessages.map((m) => m.fileId))
+    ),
+    "sId"
+  );
+
+  const contentFragmentTextByMessageId: Record<string, string[]> = {};
+  for (const m of contentFragmentMessages) {
+    if (!m.fileId || !m.contentType.startsWith("text/")) {
+      continue;
+    }
+
+    const file = contentFragmentFileBySid[m.fileId];
+    if (!file) {
+      continue;
+    }
+    const readStream = file.getReadStream({
+      auth,
+      version: "original",
+    });
+    contentFragmentTextByMessageId[m.sId] =
+      await readFirstFiveLines(readStream);
+  }
+
+  return (
+    `${visualizationSystemPrompt.trim()}\n\nYou have access to the following files:\n` +
+    contentFragmentMessages
+      .map((m) => {
+        return `<file id="${m.fileId}" name="${m.title}" type="${m.contentType}">\n${contentFragmentTextByMessageId[m.sId]?.join("\n")}(truncated...)</file>`;
+      })
+      .join("\n")
+  );
+}
+
 export const visualizationSystemPrompt = `
 You have the ability to generate visualizations (using React components) that will be rendered in the user's browser by following these instructions:
